+++ sync-for-payables/RELEASES.md
@@ -97,3 +97,13 @@ Based on:
 - [csharp v3.1.0] sync-for-payables
 ### Releases
 - [NuGet v3.1.0] https://www.nuget.org/packages/Codat.Sync.Payables/3.1.0 - sync-for-payables
+
+## 2024-02-14 12:07:04
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.179.0 (2.257.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.2.0] sync-for-payables
+### Releases
+- [NuGet v3.2.0] https://www.nuget.org/packages/Codat.Sync.Payables/3.2.0 - sync-for-payables